rearrange utils methods and remove domain_loading.py
